Relax invariants to fix #23, add tests on problematic example bits
The tests make sure that the index is actually working correctly despite
the relaxed invariant. There are much shorter possible test cases that
would have provoked the failure, but this one was discovered in real
usage and does double duty as a test on a larger vector than the other
tests.